moves load_database into sm_database
