+++ smtk/__init__.py
@@ -16,9 +16,7 @@
 # You should have received a copy of the GNU Affero General Public License
 # along with OpenQuake. If not, see <http://www.gnu.org/licenses/>.
 import os
-import pickle
 import subprocess
-from smtk.sm_database import GroundMotionDatabase
 
 
 def git_suffix(fname):
@@ -46,35 +44,6 @@ def git_suffix(fname):
             pass
 
     return ''
-
-
-def load_database(directory):
-    """
-    Wrapper function to load the metadata of a strong motion database
-    according to the filetype
-    """
-    metadata_file = None
-    filetype = None
-    fileset = os.listdir(directory)
-    for ftype in ["pkl", "json"]:
-        if ("metadatafile.%s" % ftype) in fileset:
-            metadata_file = "metadatafile.%s" % ftype
-            filetype = ftype
-            break
-    if not metadata_file:
-        raise IOError(
-            "Expected metadata file of supported type not found in %s"
-            % directory)
-    metadata_path = os.path.join(directory, metadata_file)
-    if filetype == "json":
-        # json metadata filetype
-        return GroundMotionDatabase.from_json(metadata_path)
-    elif filetype == "pkl":
-        # pkl file type
-        with open(metadata_path, "rb") as f:
-            return pickle.load(f)
-    else:
-        raise ValueError("Metadata filetype %s not supported" % ftype)
 
 
 # version is used by the setup.py
